% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deg}
\alias{deg}
\title{DEG matrix of breast cancer (part)}
\format{a matrix}
\value{
A matrix
}
\description{
A matrix of bianry DEG matrix for breast cancer patients (only part)
}
\keyword{datasets}
