% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp}
\alias{exp}
\title{expression matrix of breast cancer (part)}
\format{a matrix}
\value{
A matrix
}
\description{
A matrix of breast cancer patients from TCGA (only part)
}
\keyword{datasets}
