% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.deg.R
\name{Plot.deg}
\alias{Plot.deg}
\alias{Plot}
\title{Plot the DEGs before or after cross-validation}
\usage{
\method{Plot}{deg}(input, ann = NULL, col.order = NULL, show.genes = NULL,
  max.n = 30, up.col = "red", down.col = "blue", ...)

Plot(...)
}
\arguments{
\item{input}{a 'deg' object returned by \code{\link{bi.deg}}}

\item{ann}{a data.frame for the patient annotation}

\item{col.order}{the order of column in heatmap}

\item{show.genes}{the gene ids to plot}

\item{max.n}{the maximum number of genes to plot}

\item{up.col}{the color for up-regulated genes}

\item{down.col}{the color for down-regulated genes}

\item{...}{other setting of 'oncoPrint'}
}
\value{
A heatmap plot
}
\description{
Plot the binary differential expression matrix transformed by \code{\link{bi.deg}}
}
\details{
This function applied the function of oncoPrint from `ComplexHeatmap` to dispaly ownership of the DEGs. The output is a heatmap plots where the genes with maximum observations are showed.
}
\examples{
Plot(deg,ann.er, max.n=5)
Plot(deg.spc, ann.er, max.n=5)

}
\author{
Guofeng Meng
}
