% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.deg.R
\name{Plot.deg.specific.test}
\alias{Plot.deg.specific.test}
\title{Plot the DEGs before or after cross-validation}
\usage{
\method{Plot}{deg.specific.test}(input, ann = NULL, col.order = NULL,
  show.genes = NULL, max.n = 30, up.col = "red", down.col = "blue", ...)
}
\arguments{
\item{input}{a 'deg.specific' object returned by \code{\link{deg.specific}}}

\item{ann}{a data.frame for the patient annotation}

\item{col.order}{the order of column in heatmap}

\item{show.genes}{the gene ids to plot}

\item{max.n}{the maximum number of genes to plot}

\item{up.col}{the color for up-regulated genes}

\item{down.col}{the color for down-regulated genes}

\item{...}{other setting of 'oncoPrint'}
}
\value{
A heatmap plot
}
\description{
Plot the cross-validated DEGs predicted by \code{\link{deg.specific}}.
}
\details{
This function applied the function of oncoPrint from `ComplexHeatmap` to dispaly ownership of the DEGs. The output is a heatmap plots where the genes with maximum observations are showed.
}
\examples{
\dontrun{
Plot(deg,ann.er, max.n=5)
Plot(deg.spc, ann.er, max.n=5)
}

}
\references{

}
\author{
Guofeng Meng
}
