% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.module.R
\name{Plot.seed.module}
\alias{Plot.seed.module}
\title{Plot the DEGs modules shared by patients}
\usage{
\method{Plot}{seed.module}(res.module, ann = NULL, deg = NULL,
  col.order = NULL, show.mods = NULL, overlap = NULL,
  dissimilarity = NULL, max.n = min(length(res.module), 30),
  type = c("model", "max.patients", "max.genes")[1], label.col = "blue",
  ...)
}
\arguments{
\item{res.module}{a 'seed.module' object returned by \code{\link{seed.module}}}

\item{ann}{a data.frame for the patient annotation}

\item{deg}{a 'deg' to display. It is returned by \code{\link{bi.deg}}}

\item{col.order}{the order of column in heatmap}

\item{show.mods}{a vector, the modules to display}

\item{overlap}{the similarity cutoff to display as carrying the module}

\item{dissimilarity}{the similarity cutoff to display as not carrying the module}

\item{max.n}{the maximum number of modules to display}

\item{type}{the module type to display}

\item{label.col}{the color to label}

\item{...}{other setting of 'oncoPrint'}
}
\value{
A heatmap plot
}
\description{
Plot the DEGs modules
}
\details{
This function is to display the relationship of the predicted DEG modules and the patients.

'deg' can be set to display the modules from different datasets, e.g. the modules predicted from disease A and display them in the binary DEG matrix of disease B.

The output is a heatmap Plot where the modules with maximum observations are showed.
}
\examples{
Plot(seed.mod, ann.er, max.n=5)
Plot(seed.mod, ann.er, deg=deg, max.n=5)
}
\author{
Guofeng Meng
}
