% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.module.R
\name{summarize.cluster.module}
\alias{summarize.cluster.module}
\title{Summarize the DEG modules}
\usage{
\method{summarize}{cluster.module}(res.module, max.n = 10, ...)
}
\arguments{
\item{res.module}{a 'cluster.module' object returned by \code{\link{cluster.module}}}

\item{max.n}{the number of modules to display}

\item{...}{other setting}
}
\value{
A data.frame.
}
\description{
Summarize the DEG modules
}
\details{
This function summarize the DEG modules
}
\examples{

report.module <- summarize(cluster.mod)

}
\references{

}
\author{
Guofeng Meng
}
