% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.module.R
\name{summarize.seed.module}
\alias{summarize.seed.module}
\alias{summarize}
\title{Summarize the DEG modules}
\usage{
\method{summarize}{seed.module}(res.module, max.n = 10, ...)

summarize(...)
}
\arguments{
\item{res.module}{a 'seed.module' object returned by \code{\link{seed.module}}}

\item{max.n}{the number of modules to display}

\item{...}{other setting}
}
\value{
A data.frame.
}
\description{
Summarize the DEG modules
}
\details{
This function summarize the DEG modules
}
\examples{
report.module <- summarize(seed.mod)
}
\references{

}
\author{
Guofeng Meng
}
